Updated the scheduler, added an integrity check to db, adjusted JSON processing, improved SQL to fetch pre-formed dataframe, updated streamlit version.
