+++ requirements.txt
@@ -8,6 +8,6 @@ pystan==2.19.1.1
 python-dotenv==0.15.0
 requests==2.25.0
 scikit_learn==0.23.2
-seaborn==0.11.0
+seaborn
 SQLAlchemy==1.3.20
-streamlit==0.72.0 
+streamlit
