+++ static/data/accuracy_dataframe.csv
@@ -1,64 +1,98 @@
-,today+0,today+1,today+2,today+3,today+4,today+5,today+6
-2020-10-08,0,-1.0,-1.0,6.0,-7.0,-2.0,-2.0
-2020-10-09,0,0.0,2.0,-9.0,-3.0,-2.0,0.0
-2020-10-10,0,0.0,-5.0,1.0,2.0,0.0,2.0
-2020-10-11,0,-3.0,-2.0,-4.0,0.0,2.0,-4.0
-2020-10-12,0,-1.0,-1.0,-1.0,-1.0,-4.0,-1.0
-2020-10-13,0,0.0,0.0,-1.0,-3.0,0.0,0.0
-2020-10-14,0,0.0,-1.0,-3.0,0.0,0.0,-4.0
-2020-10-15,0,-1.0,-3.0,0.0,0.0,-4.0,-5.0
-2020-10-16,0,0.0,-1.0,-2.0,-4.0,-5.0,5.0
-2020-10-17,0,0.0,-2.0,-2.0,-3.0,2.0,-7.0
-2020-10-18,0,-1.0,0.0,2.0,2.0,0.0,-5.0
-2020-10-19,0,-1.0,1.0,3.0,0.0,-2.0,1.0
-2020-10-20,0,0.0,0.0,0.0,-1.0,0.0,-2.0
-2020-10-21,0,0.0,0.0,-1.0,0.0,-2.0,1.0
-2020-10-22,0,0.0,-1.0,0.0,-2.0,1.0,2.0
-2020-10-23,0,-1.0,0.0,-2.0,1.0,1.0,0.0
-2020-10-24,0,0.0,0.0,1.0,1.0,2.0,0.0
-2020-10-25,0,1.0,2.0,2.0,2.0,0.0,-1.0
-2020-10-26,0,1.0,2.0,2.0,1.0,-2.0,2.0
-2020-10-27,0,0.0,0.0,0.0,-1.0,1.0,-2.0
-2020-10-28,0,0.0,0.0,-1.0,-1.0,-3.0,9.0
-2020-10-29,0,0.0,-1.0,-1.0,-3.0,9.0,7.0
-2020-10-30,0,0.0,0.0,-4.0,6.0,8.0,-2.0
-2020-10-31,0,0.0,-4.0,6.0,8.0,-2.0,0.0
-2020-11-01,0,-4.0,6.0,8.0,-2.0,0.0,0.0
-2020-11-02,0,1.0,-1.0,0.0,-2.0,0.0,2.0
-2020-11-03,0,0.0,0.0,0.0,1.0,2.0,5.0
-2020-11-04,0,0.0,0.0,0.0,0.0,3.0,2.0
-2020-11-05,0,0.0,0.0,0.0,2.0,2.0,6.0
-2020-11-06,0,0.0,0.0,2.0,2.0,5.0,8.0
-2020-11-07,0,0.0,2.0,0.0,5.0,9.0,6.0
-2020-11-08,0,2.0,0.0,2.0,5.0,4.0,3.0
-2020-11-09,0,0.0,1.0,5.0,5.0,1.0,-7.0
-2020-11-10,0,1.0,5.0,5.0,2.0,2.0,14.0
-2020-11-11,0,5.0,5.0,2.0,-10.0,6.0,3.0
-2020-11-12,0,4.0,2.0,-10.0,13.0,3.0,-3.0
-2020-11-13,0,1.0,-12.0,10.0,4.0,-3.0,-3.0
-2020-11-14,0,-12.0,10.0,3.0,-5.0,-3.0,15.0
-2020-11-15,0,10.0,3.0,-6.0,-4.0,13.0,-1.0
-2020-11-16,0,3.0,-7.0,-5.0,16.0,-2.0,4.0
-2020-11-17,0,-7.0,-5.0,16.0,-2.0,1.0,-2.0
-2020-11-18,0,-5.0,13.0,-2.0,1.0,-2.0,0.0
-2020-11-19,0,13.0,-5.0,0.0,0.0,0.0,-9.0
-2020-11-20,0,-5.0,0.0,0.0,0.0,-9.0,-11.0
-2020-11-21,0,-4.0,0.0,0.0,-9.0,-9.0,3.0
-2020-11-22,0,0.0,-1.0,-9.0,-4.0,7.0,0.0
-2020-11-23,0,0.0,-9.0,-2.0,3.0,-11.0,2.0
-2020-11-24,0,-10.0,-1.0,2.0,-4.0,3.0,-3.0
-2020-11-25,0,-1.0,3.0,-3.0,3.0,-4.0,2.0
-2020-11-26,0,1.0,-3.0,5.0,-6.0,-5.0,8.0
-2020-11-27,0,-5.0,5.0,-6.0,-4.0,8.0,-3.0
-2020-11-28,0,6.0,-6.0,-3.0,8.0,-3.0,0.0
-2020-11-29,0,-7.0,-3.0,8.0,-3.0,0.0,-1.0
-2020-11-30,0,-3.0,8.0,-3.0,0.0,2.0,-4.0
-2020-12-01,0,8.0,-3.0,-1.0,4.0,0.0,-2.0
-2020-12-02,0,-4.0,-1.0,-2.0,2.0,1.0,-1.0
-2020-12-03,0,-1.0,0.0,0.0,-1.0,0.0,4.0
-2020-12-04,0,2.0,0.0,-1.0,0.0,3.0,
-2020-12-05,0,0.0,-1.0,-1.0,3.0,,
-2020-12-06,0,0.0,0.0,2.0,,,
-2020-12-07,0,0.0,1.0,,,,
-2020-12-08,0,0.0,,,,,
-2020-12-09,0,,,,,,
+issue,today+0,today+1,today+2,today+3,today+4,today+5,today+6
+2020-10-08,0.0,-1.0,-1.0,6.0,-7.0,-2.0,-2.0
+2020-10-09,0.0,0.0,2.0,-9.0,-3.0,-2.0,0.0
+2020-10-10,0.0,0.0,-5.0,1.0,2.0,0.0,2.0
+2020-10-11,0.0,-3.0,-2.0,-4.0,0.0,2.0,-4.0
+2020-10-12,0.0,-1.0,-1.0,-1.0,-1.0,-4.0,-1.0
+2020-10-13,0.0,0.0,0.0,-1.0,-3.0,0.0,0.0
+2020-10-14,0.0,0.0,-1.0,-3.0,0.0,0.0,-4.0
+2020-10-15,0.0,-1.0,-3.0,0.0,0.0,-4.0,-5.0
+2020-10-16,0.0,0.0,-1.0,-2.0,-4.0,-5.0,0.0
+2020-10-17,0.0,0.0,-2.0,-2.0,-3.0,2.0,-2.0
+2020-10-18,0.0,-1.0,0.0,2.0,2.0,0.0,-3.0
+2020-10-19,0.0,-1.0,1.0,3.0,0.0,-2.0,-1.0
+2020-10-20,0.0,0.0,0.0,0.0,-1.0,0.0,-2.0
+2020-10-21,0.0,0.0,0.0,-1.0,0.0,-2.0,1.0
+2020-10-22,0.0,0.0,-1.0,0.0,-2.0,1.0,1.0
+2020-10-23,0.0,-1.0,0.0,-2.0,1.0,1.0,2.0
+2020-10-24,0.0,0.0,0.0,1.0,1.0,2.0,-1.0
+2020-10-25,0.0,1.0,2.0,2.0,2.0,0.0,-1.0
+2020-10-26,0.0,1.0,2.0,2.0,1.0,-2.0,2.0
+2020-10-27,0.0,0.0,0.0,0.0,-1.0,1.0,-2.0
+2020-10-28,0.0,0.0,0.0,-1.0,-1.0,-3.0,9.0
+2020-10-29,0.0,0.0,-1.0,-1.0,-3.0,9.0,7.0
+2020-10-30,0.0,0.0,0.0,-4.0,6.0,8.0,-2.0
+2020-10-31,0.0,0.0,-4.0,6.0,8.0,-2.0,0.0
+2020-11-01,0.0,-4.0,6.0,8.0,-2.0,0.0,0.0
+2020-11-02,0.0,1.0,-1.0,0.0,-2.0,0.0,2.0
+2020-11-03,0.0,0.0,0.0,0.0,1.0,2.0,5.0
+2020-11-04,0.0,0.0,0.0,0.0,0.0,3.0,2.0
+2020-11-05,0.0,0.0,0.0,0.0,2.0,2.0,4.0
+2020-11-06,0.0,0.0,0.0,2.0,2.0,3.0,4.0
+2020-11-07,0.0,0.0,2.0,0.0,3.0,4.0,3.0
+2020-11-08,0.0,2.0,0.0,0.0,1.0,1.0,3.0
+2020-11-09,0.0,0.0,-1.0,1.0,2.0,2.0,5.0
+2020-11-10,0.0,-1.0,1.0,2.0,2.0,6.0,4.0
+2020-11-11,0.0,1.0,2.0,2.0,2.0,3.0,0.0
+2020-11-12,0.0,1.0,2.0,2.0,3.0,1.0,5.0
+2020-11-13,0.0,1.0,0.0,0.0,1.0,5.0,3.0
+2020-11-14,0.0,0.0,0.0,0.0,3.0,3.0,2.0
+2020-11-15,0.0,0.0,0.0,2.0,2.0,0.0,4.0
+2020-11-16,0.0,0.0,1.0,1.0,3.0,3.0,7.0
+2020-11-17,0.0,1.0,1.0,3.0,3.0,4.0,-2.0
+2020-11-18,0.0,1.0,0.0,3.0,4.0,-2.0,0.0
+2020-11-19,0.0,0.0,0.0,3.0,0.0,0.0,1.0
+2020-11-20,0.0,0.0,3.0,0.0,0.0,1.0,-11.0
+2020-11-21,0.0,-1.0,0.0,0.0,1.0,-9.0,3.0
+2020-11-22,0.0,0.0,-1.0,1.0,-4.0,7.0,2.0
+2020-11-23,0.0,0.0,1.0,-2.0,3.0,-9.0,-4.0
+2020-11-24,0.0,0.0,-1.0,2.0,-2.0,-3.0,3.0
+2020-11-25,0.0,-1.0,3.0,-1.0,-3.0,2.0,4.0
+2020-11-26,0.0,1.0,-1.0,-1.0,0.0,-3.0,0.0
+2020-11-27,0.0,-3.0,-1.0,0.0,-2.0,0.0,1.0
+2020-11-28,0.0,0.0,0.0,-1.0,0.0,1.0,0.0
+2020-11-29,0.0,-1.0,-1.0,0.0,1.0,0.0,-1.0
+2020-11-30,0.0,-1.0,0.0,1.0,0.0,2.0,-4.0
+2020-12-01,0.0,0.0,1.0,-1.0,4.0,0.0,-2.0
+2020-12-02,0.0,0.0,-1.0,-2.0,2.0,1.0,-1.0
+2020-12-03,0.0,-1.0,0.0,0.0,-1.0,0.0,4.0
+2020-12-04,0.0,2.0,0.0,-1.0,0.0,3.0,-4.0
+2020-12-05,0.0,0.0,-1.0,-1.0,3.0,-4.0,0.0
+2020-12-06,0.0,0.0,0.0,2.0,-3.0,0.0,2.0
+2020-12-07,0.0,0.0,1.0,0.0,2.0,2.0,-2.0
+2020-12-08,0.0,0.0,0.0,1.0,2.0,1.0,2.0
+2020-12-09,0.0,0.0,0.0,2.0,1.0,3.0,-3.0
+2020-12-10,0.0,2.0,1.0,0.0,1.0,-3.0,4.0
+2020-12-11,0.0,2.0,0.0,2.0,-3.0,4.0,4.0
+2020-12-12,0.0,0.0,2.0,-2.0,2.0,2.0,-1.0
+2020-12-13,0.0,2.0,0.0,3.0,1.0,-3.0,-3.0
+2020-12-14,0.0,-1.0,3.0,4.0,0.0,1.0,-3.0
+2020-12-15,0.0,2.0,1.0,1.0,-1.0,-2.0,0.0
+2020-12-16,0.0,-1.0,0.0,-1.0,-2.0,1.0,-1.0
+2020-12-17,0.0,0.0,0.0,-2.0,-4.0,-2.0,-1.0
+2020-12-18,0.0,0.0,-1.0,-1.0,-3.0,-1.0,-4.0
+2020-12-19,0.0,0.0,-1.0,-3.0,-2.0,-3.0,-1.0
+2020-12-20,0.0,-2.0,-1.0,0.0,-2.0,-1.0,0.0
+2020-12-21,0.0,0.0,0.0,-1.0,-1.0,-1.0,7.0
+2020-12-22,0.0,0.0,-1.0,-1.0,-1.0,7.0,-2.0
+2020-12-23,0.0,0.0,-1.0,-3.0,5.0,-1.0,-1.0
+2020-12-24,0.0,0.0,0.0,4.0,-1.0,1.0,-1.0
+2020-12-25,0.0,0.0,3.0,-1.0,0.0,-1.0,-1.0
+2020-12-26,0.0,1.0,-1.0,0.0,0.0,-1.0,-1.0
+2020-12-27,0.0,0.0,0.0,0.0,-2.0,-1.0,-1.0
+2020-12-28,0.0,0.0,-1.0,-2.0,-2.0,0.0,-2.0
+2020-12-29,0.0,0.0,0.0,-1.0,1.0,-2.0,-4.0
+2020-12-30,0.0,1.0,-2.0,0.0,-1.0,-5.0,-3.0
+2020-12-31,0.0,-1.0,0.0,-1.0,-4.0,-4.0,-2.0
+2021-01-01,0.0,1.0,-1.0,-3.0,-2.0,-2.0,-1.0
+2021-01-02,0.0,0.0,-2.0,-2.0,-2.0,-1.0,-3.0
+2021-01-03,0.0,-2.0,-1.0,-1.0,-1.0,-3.0,-4.0
+2021-01-04,0.0,-1.0,-1.0,0.0,-1.0,0.0,7.0
+2021-01-05,0.0,-1.0,1.0,0.0,1.0,4.0,3.0
+2021-01-06,0.0,1.0,0.0,0.0,3.0,3.0,-5.0
+2021-01-07,0.0,1.0,1.0,2.0,1.0,-5.0,
+2021-01-08,0.0,-1.0,2.0,1.0,-2.0,,
+2021-01-09,0.0,0.0,1.0,-1.0,,,
+2021-01-10,0.0,1.0,-1.0,,,,
+2021-01-11,0.0,-1.0,,,,,
+2021-01-12,0.0,,,,,,
