+++ static/data/forecast_dataframe.csv
@@ -1,64 +1,98 @@
-,today+0,today+1,today+2,today+3,today+4,today+5,today+6
-2020-10-08,18,22,18,18,23,26,27.0
-2020-10-09,21,17,22,25,27,27,19.0
-2020-10-10,17,24,21,23,23,19,18.0
-2020-10-11,24,19,26,29,19,18,19.0
-2020-10-12,16,25,26,20,21,19,17.0
-2020-10-13,24,25,19,21,18,16,17.0
-2020-10-14,25,19,21,18,16,17,22.0
-2020-10-15,19,21,18,16,17,22,25.0
-2020-10-16,20,15,17,19,22,25,18.0
-2020-10-17,15,16,19,20,23,21,23.0
-2020-10-18,16,18,18,18,21,16,18.0
-2020-10-19,17,19,19,20,16,15,16.0
-2020-10-20,18,20,23,16,14,17,20.0
-2020-10-21,20,23,16,14,17,20,19.0
-2020-10-22,23,16,14,17,20,19,19.0
-2020-10-23,16,14,17,20,19,20,20.0
-2020-10-24,13,17,18,19,20,18,18.0
-2020-10-25,17,17,18,19,18,18,19.0
-2020-10-26,18,19,19,18,17,20,23.0
-2020-10-27,20,21,20,18,19,24,27.0
-2020-10-28,21,20,18,19,26,28,21.0
-2020-10-29,20,18,19,26,28,21,16.0
-2020-10-30,18,18,25,29,24,15,17.0
-2020-10-31,18,25,29,24,15,17,16.0
-2020-11-01,25,29,24,15,17,16,15.0
-2020-11-02,25,29,24,15,18,15,18.0
-2020-11-03,30,23,15,16,14,18,23.0
-2020-11-04,23,15,16,15,20,25,28.0
-2020-11-05,15,16,15,20,26,28,24.0
-2020-11-06,16,15,20,26,28,25,20.0
-2020-11-07,15,20,26,30,25,19,18.0
-2020-11-08,20,26,30,28,23,20,18.0
-2020-11-09,28,30,29,23,19,20,28.0
-2020-11-10,30,29,23,19,19,19,19.0
-2020-11-11,30,23,19,19,31,27,20.0
-2020-11-12,28,20,19,31,20,20,23.0
-2020-11-13,24,20,33,23,19,23,31.0
-2020-11-14,21,33,23,20,25,31,19.0
-2020-11-15,21,23,20,26,32,21,22.0
-2020-11-16,33,20,27,33,18,23,22.0
-2020-11-17,23,27,33,18,23,25,22.0
-2020-11-18,20,33,21,23,25,22,20.0
-2020-11-19,28,21,26,26,20,20,29.0
-2020-11-20,34,26,26,20,20,29,31.0
-2020-11-21,21,30,20,20,29,29,31.0
-2020-11-22,26,20,21,29,24,27,23.0
-2020-11-23,20,20,29,22,31,34,23.0
-2020-11-24,20,30,21,32,27,22,22.0
-2020-11-25,20,21,31,26,22,23,23.0
-2020-11-26,20,33,26,20,25,30,19.0
-2020-11-27,34,28,20,25,29,19,22.0
-2020-11-28,23,19,25,28,19,22,23.0
-2020-11-29,25,26,28,19,22,23,29.0
-2020-11-30,19,28,19,22,23,26,24.0
-2020-12-01,25,19,22,24,24,20,20.0
-2020-12-02,27,23,24,30,18,17,19.0
-2020-12-03,19,24,28,20,19,18,21.0
-2020-12-04,23,26,20,19,18,22,22.0
-2020-12-05,28,20,19,19,22,22,32.0
-2020-12-06,20,18,18,23,21,23,33.0
-2020-12-07,18,18,24,18,21,27,
-2020-12-08,18,25,18,22,27,29,
-2020-12-09,25,18,23,27,29,31,
+issue,today+0,today+1,today+2,today+3,today+4,today+5,today+6
+2020-10-08,18,22,18,18,23,26,27
+2020-10-09,21,17,22,25,27,27,19
+2020-10-10,17,24,21,23,23,19,18
+2020-10-11,24,19,26,29,19,18,19
+2020-10-12,16,25,26,20,21,19,17
+2020-10-13,24,25,19,21,18,16,17
+2020-10-14,25,19,21,18,16,17,22
+2020-10-15,19,21,18,16,17,22,25
+2020-10-16,20,15,17,19,22,25,23
+2020-10-17,15,16,19,20,23,21,18
+2020-10-18,16,18,18,18,21,16,16
+2020-10-19,17,19,19,20,16,15,18
+2020-10-20,18,20,23,16,14,17,20
+2020-10-21,20,23,16,14,17,20,19
+2020-10-22,23,16,14,17,20,19,20
+2020-10-23,16,14,17,20,19,20,18
+2020-10-24,13,17,18,19,20,18,19
+2020-10-25,17,17,18,19,18,18,19
+2020-10-26,18,19,19,18,17,20,23
+2020-10-27,20,21,20,18,19,24,27
+2020-10-28,21,20,18,19,26,28,21
+2020-10-29,20,18,19,26,28,21,16
+2020-10-30,18,18,25,29,24,15,17
+2020-10-31,18,25,29,24,15,17,16
+2020-11-01,25,29,24,15,17,16,15
+2020-11-02,25,29,24,15,18,15,18
+2020-11-03,30,23,15,16,14,18,23
+2020-11-04,23,15,16,15,20,25,28
+2020-11-05,15,16,15,20,26,28,24
+2020-11-06,16,15,20,26,28,25,20
+2020-11-07,15,20,26,30,25,20,18
+2020-11-08,20,26,30,28,23,20,18
+2020-11-09,28,30,29,23,19,19,28
+2020-11-10,30,29,23,19,19,27,19
+2020-11-11,28,23,19,19,31,20,20
+2020-11-12,24,20,19,31,20,19,23
+2020-11-13,21,20,33,23,19,23,31
+2020-11-14,21,33,23,20,25,31,19
+2020-11-15,33,23,20,26,32,21,22
+2020-11-16,23,20,27,33,18,23,22
+2020-11-17,20,27,33,18,23,25,22
+2020-11-18,28,33,21,23,25,22,20
+2020-11-19,34,21,26,26,20,20,29
+2020-11-20,21,26,26,20,20,29,31
+2020-11-21,26,30,20,20,29,29,31
+2020-11-22,29,20,21,29,24,27,23
+2020-11-23,20,20,29,22,31,34,23
+2020-11-24,20,30,21,32,27,22,22
+2020-11-25,30,21,31,26,22,23,23
+2020-11-26,20,33,26,20,25,30,19
+2020-11-27,34,28,20,25,29,19,22
+2020-11-28,25,19,25,28,19,22,23
+2020-11-29,19,26,28,19,22,23,29
+2020-11-30,25,28,19,22,23,26,24
+2020-12-01,27,19,22,24,24,20,20
+2020-12-02,19,23,24,30,18,17,19
+2020-12-03,23,24,28,20,19,18,21
+2020-12-04,23,26,20,19,18,22,22
+2020-12-05,28,20,19,19,22,22,23
+2020-12-06,20,18,18,23,21,23,27
+2020-12-07,18,18,24,18,21,27,32
+2020-12-08,18,25,18,22,27,29,32
+2020-12-09,25,18,23,27,29,31,33
+2020-12-10,18,21,28,30,33,33,22
+2020-12-11,23,27,30,32,33,22,23
+2020-12-12,29,30,32,32,24,25,20
+2020-12-13,30,32,30,23,26,22,23
+2020-12-14,34,31,23,23,19,19,23
+2020-12-15,30,24,26,18,21,22,23
+2020-12-16,26,28,19,21,22,22,19
+2020-12-17,27,19,20,22,27,20,21
+2020-12-18,19,20,21,24,21,21,23
+2020-12-19,20,20,24,21,22,22,21
+2020-12-20,20,25,19,20,21,21,27
+2020-12-21,23,18,20,20,21,28,27
+2020-12-22,18,20,20,21,28,27,20
+2020-12-23,20,19,21,30,29,19,23
+2020-12-24,19,20,27,30,19,21,24
+2020-12-25,20,27,31,19,22,24,24
+2020-12-26,27,33,19,22,23,24,27
+2020-12-27,34,18,22,23,25,27,28
+2020-12-28,18,22,24,25,28,27,26
+2020-12-29,22,23,23,27,26,26,23
+2020-12-30,23,22,28,27,25,24,23
+2020-12-31,23,27,27,25,23,24,21
+2021-01-01,26,26,25,22,22,21,27
+2021-01-02,27,24,21,22,21,27,29
+2021-01-03,24,21,21,20,27,29,31
+2021-01-04,19,21,20,26,27,27,25
+2021-01-05,20,20,25,26,26,28,35
+2021-01-06,19,25,26,27,29,35,27
+2021-01-07,26,25,26,30,37,27,31
+2021-01-08,26,28,30,37,24,32,23
+2021-01-09,27,32,37,23,31,22,21
+2021-01-10,32,37,23,31,22,21,19
+2021-01-11,38,23,34,21,21,19,22
+2021-01-12,22,33,21,20,21,22,21
