+++ static/data/last_row.csv
@@ -1,2 +1,1 @@
 date,date,forecast,temp_max,issue,extended_text
-2020-12-14 13:00:00+00:00,2020-12-14T13:00:00Z,6,33,2020-12-09,"Mostly sunny morning. Medium (40%) chance of showers, most likely later in the day. Winds northerly 20 to 30 km/h shifting south to southwesterly 15 to 25 km/h during the day."
