+++ static/data/persistence_dataframe.csv
@@ -1,4 +1,4 @@
-,Persistence Accuracy,1 Day Forecast,2 Day Forecast,3 Day Forecast,4 Day Forecast,5 Day Forecast,6 Day Forecast
+issue,Persistence Accuracy,1 Day Forecast,2 Day Forecast,3 Day Forecast,4 Day Forecast,5 Day Forecast,6 Day Forecast
 2020-10-08,-3,-1.0,-1.0,6.0,-7.0,-2.0,-2.0
 2020-10-09,4,0.0,2.0,-9.0,-3.0,-2.0,0.0
 2020-10-10,-7,0.0,-5.0,1.0,2.0,0.0,2.0
@@ -7,15 +7,15 @@
 2020-10-13,-1,0.0,0.0,-1.0,-3.0,0.0,0.0
 2020-10-14,6,0.0,-1.0,-3.0,0.0,0.0,-4.0
 2020-10-15,-1,-1.0,-3.0,0.0,0.0,-4.0,-5.0
-2020-10-16,5,0.0,-1.0,-2.0,-4.0,-5.0,5.0
-2020-10-17,-1,0.0,-2.0,-2.0,-3.0,2.0,-7.0
-2020-10-18,-1,-1.0,0.0,2.0,2.0,0.0,-5.0
-2020-10-19,-1,-1.0,1.0,3.0,0.0,-2.0,1.0
+2020-10-16,5,0.0,-1.0,-2.0,-4.0,-5.0,0.0
+2020-10-17,-1,0.0,-2.0,-2.0,-3.0,2.0,-2.0
+2020-10-18,-1,-1.0,0.0,2.0,2.0,0.0,-3.0
+2020-10-19,-1,-1.0,1.0,3.0,0.0,-2.0,-1.0
 2020-10-20,-2,0.0,0.0,0.0,-1.0,0.0,-2.0
 2020-10-21,-3,0.0,0.0,-1.0,0.0,-2.0,1.0
-2020-10-22,7,0.0,-1.0,0.0,-2.0,1.0,2.0
-2020-10-23,3,-1.0,0.0,-2.0,1.0,1.0,0.0
-2020-10-24,-4,0.0,0.0,1.0,1.0,2.0,0.0
+2020-10-22,7,0.0,-1.0,0.0,-2.0,1.0,1.0
+2020-10-23,3,-1.0,0.0,-2.0,1.0,1.0,2.0
+2020-10-24,-4,0.0,0.0,1.0,1.0,2.0,-1.0
 2020-10-25,-1,1.0,2.0,2.0,2.0,0.0,-1.0
 2020-10-26,-2,1.0,2.0,2.0,1.0,-2.0,2.0
 2020-10-27,-1,0.0,0.0,0.0,-1.0,1.0,-2.0
@@ -27,37 +27,71 @@
 2020-11-02,-5,1.0,-1.0,0.0,-2.0,0.0,2.0
 2020-11-03,7,0.0,0.0,0.0,1.0,2.0,5.0
 2020-11-04,8,0.0,0.0,0.0,0.0,3.0,2.0
-2020-11-05,-1,0.0,0.0,0.0,2.0,2.0,6.0
-2020-11-06,1,0.0,0.0,2.0,2.0,5.0,8.0
-2020-11-07,-5,0.0,2.0,0.0,5.0,9.0,6.0
-2020-11-08,-8,2.0,0.0,2.0,5.0,4.0,3.0
-2020-11-09,-2,0.0,1.0,5.0,5.0,1.0,-7.0
-2020-11-10,0,1.0,5.0,5.0,2.0,2.0,14.0
-2020-11-11,2,5.0,5.0,2.0,-10.0,6.0,3.0
-2020-11-12,4,4.0,2.0,-10.0,13.0,3.0,-3.0
-2020-11-13,3,1.0,-12.0,10.0,4.0,-3.0,-3.0
-2020-11-14,0,-12.0,10.0,3.0,-5.0,-3.0,15.0
-2020-11-15,-12,10.0,3.0,-6.0,-4.0,13.0,-1.0
-2020-11-16,10,3.0,-7.0,-5.0,16.0,-2.0,4.0
-2020-11-17,3,-7.0,-5.0,16.0,-2.0,1.0,-2.0
-2020-11-18,-8,-5.0,13.0,-2.0,1.0,-2.0,0.0
-2020-11-19,-6,13.0,-5.0,0.0,0.0,0.0,-9.0
-2020-11-20,13,-5.0,0.0,0.0,0.0,-9.0,-11.0
-2020-11-21,-5,-4.0,0.0,0.0,-9.0,-9.0,3.0
-2020-11-22,6,0.0,-1.0,-9.0,-4.0,7.0,0.0
-2020-11-23,0,0.0,-9.0,-2.0,3.0,-11.0,2.0
-2020-11-24,0,-10.0,-1.0,2.0,-4.0,3.0,-3.0
-2020-11-25,0,-1.0,3.0,-3.0,3.0,-4.0,2.0
-2020-11-26,-14,1.0,-3.0,5.0,-6.0,-5.0,8.0
-2020-11-27,11,-5.0,5.0,-6.0,-4.0,8.0,-3.0
-2020-11-28,-2,6.0,-6.0,-3.0,8.0,-3.0,0.0
-2020-11-29,6,-7.0,-3.0,8.0,-3.0,0.0,-1.0
-2020-11-30,-6,-3.0,8.0,-3.0,0.0,2.0,-4.0
-2020-12-01,-2,8.0,-3.0,-1.0,4.0,0.0,-2.0
-2020-12-02,8,-4.0,-1.0,-2.0,2.0,1.0,-1.0
-2020-12-03,-4,-1.0,0.0,0.0,-1.0,0.0,4.0
-2020-12-04,-5,2.0,0.0,-1.0,0.0,3.0,
-2020-12-05,8,0.0,-1.0,-1.0,3.0,,
-2020-12-06,2,0.0,0.0,2.0,,,
-2020-12-07,0,0.0,1.0,,,,
-2020-12-08,-7,0.0,,,,,
+2020-11-05,-1,0.0,0.0,0.0,2.0,2.0,4.0
+2020-11-06,1,0.0,0.0,2.0,2.0,3.0,4.0
+2020-11-07,-5,0.0,2.0,0.0,3.0,4.0,3.0
+2020-11-08,-8,2.0,0.0,0.0,1.0,1.0,3.0
+2020-11-09,-2,0.0,-1.0,1.0,2.0,2.0,5.0
+2020-11-10,2,-1.0,1.0,2.0,2.0,6.0,4.0
+2020-11-11,4,1.0,2.0,2.0,2.0,3.0,0.0
+2020-11-12,3,1.0,2.0,2.0,3.0,1.0,5.0
+2020-11-13,0,1.0,0.0,0.0,1.0,5.0,3.0
+2020-11-14,-12,0.0,0.0,0.0,3.0,3.0,2.0
+2020-11-15,10,0.0,0.0,2.0,2.0,0.0,4.0
+2020-11-16,3,0.0,1.0,1.0,3.0,3.0,7.0
+2020-11-17,-8,1.0,1.0,3.0,3.0,4.0,-2.0
+2020-11-18,-6,1.0,0.0,3.0,4.0,-2.0,0.0
+2020-11-19,13,0.0,0.0,3.0,0.0,0.0,1.0
+2020-11-20,-5,0.0,3.0,0.0,0.0,1.0,-11.0
+2020-11-21,-3,-1.0,0.0,0.0,1.0,-9.0,3.0
+2020-11-22,9,0.0,-1.0,1.0,-4.0,7.0,2.0
+2020-11-23,0,0.0,1.0,-2.0,3.0,-9.0,-4.0
+2020-11-24,-10,0.0,-1.0,2.0,-2.0,-3.0,3.0
+2020-11-25,10,-1.0,3.0,-1.0,-3.0,2.0,4.0
+2020-11-26,-14,1.0,-1.0,-1.0,0.0,-3.0,0.0
+2020-11-27,9,-3.0,-1.0,0.0,-2.0,0.0,1.0
+2020-11-28,6,0.0,0.0,-1.0,0.0,1.0,0.0
+2020-11-29,-6,-1.0,-1.0,0.0,1.0,0.0,-1.0
+2020-11-30,-2,-1.0,0.0,1.0,0.0,2.0,-4.0
+2020-12-01,8,0.0,1.0,-1.0,4.0,0.0,-2.0
+2020-12-02,-4,0.0,-1.0,-2.0,2.0,1.0,-1.0
+2020-12-03,0,-1.0,0.0,0.0,-1.0,0.0,4.0
+2020-12-04,-5,2.0,0.0,-1.0,0.0,3.0,-4.0
+2020-12-05,8,0.0,-1.0,-1.0,3.0,-4.0,0.0
+2020-12-06,2,0.0,0.0,2.0,-3.0,0.0,2.0
+2020-12-07,0,0.0,1.0,0.0,2.0,2.0,-2.0
+2020-12-08,-7,0.0,0.0,1.0,2.0,1.0,2.0
+2020-12-09,7,0.0,0.0,2.0,1.0,3.0,-3.0
+2020-12-10,-5,2.0,1.0,0.0,1.0,-3.0,4.0
+2020-12-11,-6,2.0,0.0,2.0,-3.0,4.0,4.0
+2020-12-12,-1,0.0,2.0,-2.0,2.0,2.0,-1.0
+2020-12-13,-4,2.0,0.0,3.0,1.0,-3.0,-3.0
+2020-12-14,4,-1.0,3.0,4.0,0.0,1.0,-3.0
+2020-12-15,4,2.0,1.0,1.0,-1.0,-2.0,0.0
+2020-12-16,-1,-1.0,0.0,-1.0,-2.0,1.0,-1.0
+2020-12-17,8,0.0,0.0,-2.0,-4.0,-2.0,-1.0
+2020-12-18,-1,0.0,-1.0,-1.0,-3.0,-1.0,-4.0
+2020-12-19,0,0.0,-1.0,-3.0,-2.0,-3.0,-1.0
+2020-12-20,-3,-2.0,-1.0,0.0,-2.0,-1.0,0.0
+2020-12-21,5,0.0,0.0,-1.0,-1.0,-1.0,7.0
+2020-12-22,-2,0.0,-1.0,-1.0,-1.0,7.0,-2.0
+2020-12-23,1,0.0,-1.0,-3.0,5.0,-1.0,-1.0
+2020-12-24,-1,0.0,0.0,4.0,-1.0,1.0,-1.0
+2020-12-25,-7,0.0,3.0,-1.0,0.0,-1.0,-1.0
+2020-12-26,-7,1.0,-1.0,0.0,0.0,-1.0,-1.0
+2020-12-27,16,0.0,0.0,0.0,-2.0,-1.0,-1.0
+2020-12-28,-4,0.0,-1.0,-2.0,-2.0,0.0,-2.0
+2020-12-29,-1,0.0,0.0,-1.0,1.0,-2.0,-4.0
+2020-12-30,0,1.0,-2.0,0.0,-1.0,-5.0,-3.0
+2020-12-31,-3,-1.0,0.0,-1.0,-4.0,-4.0,-2.0
+2021-01-01,-1,1.0,-1.0,-3.0,-2.0,-2.0,-1.0
+2021-01-02,3,0.0,-2.0,-2.0,-2.0,-1.0,-3.0
+2021-01-03,5,-2.0,-1.0,-1.0,-1.0,-3.0,-4.0
+2021-01-04,-1,-1.0,-1.0,0.0,-1.0,0.0,7.0
+2021-01-05,1,-1.0,1.0,0.0,1.0,4.0,3.0
+2021-01-06,-7,1.0,0.0,0.0,3.0,3.0,-5.0
+2021-01-07,0,1.0,1.0,2.0,1.0,-5.0,
+2021-01-08,-1,-1.0,2.0,1.0,-2.0,,
+2021-01-09,-5,0.0,1.0,-1.0,,,
+2021-01-10,-6,1.0,-1.0,,,,
+2021-01-11,16,-1.0,,,,,
